requirements.txt
python-dotenv==0.18.0 
watchdog==2.1.3
blinker==1.4
Flask==2.0.1
Flask-Cors==3.0.10
# neo4j support
py2neo==2021.1.5
# graphql support
graphene==2.1.9
Flask-GraphQL==2.0.1
# postgres support
psycopg2-binary==2.9.1
Flask-SQLAlchemy==2.5.1
flask-restx==0.5.0